Step4 - 자동차 경주(우승자) (#4473)
* docs(README.md): 기능 목록 갱신

- step4 기능 요구사항에 대한 기능 목록 갱신

* feat(MoveStrategy): Car 객체 이동 여부 판단용 전략 객체 신규

- MoveStrategy: 이동 전략 인터페이스
- RandomMoveStrategy: MoveStrategy 구현체(랜덤하게 생성된 정수를 활용해 이동 여부 판단)

* refactor(Car): Car 객체에 이동 전략 인스턴스 변수 추가

- 이동 전략 인스턴스 변수 추가
- 이동 여부 판단을 위한 랜던 값 생성기 및 기준치 인스턴스 변수 삭제

* refactor(RaceMonitorTest): UI 관련 단위 테스크 코드 삭제

삭제 코드
- RaceMonitorTest: 자동차 위치 출력 테스트

* refactor(Car): 자동차 현재 위치 빌더 메소드 신규

- currentPosition(): 자동차 현재 위치 빌더 메소드

* feat(ConsoleInputUtil): 자동차 이름 콘솔 입력 받기

- csv 형태로 자동차 이름 콘솔 입력 받기
- 자동차 이름 5자 초과 시, 예외 발생
- 자동차 대수 콘솔 입력 메소드는 미사용으로 deprecated

* refactor(ConsoleInputUtil): 이름 길이 검증 책임 변경(presentation → domain)

- 이름 길이 검증 책임 변경: ConsoleInputUtil → Car
- 이름 길이 예외 레이어 변경: presentation → domain

* feat(Car): 자동차 이동 횟수 반환 메소드 신규

- 자동차 이동 횟수 반환 메소드 신규: numberOfMove()

* refactor(ConsoleInputUtil): 자동차 대수, 시도 횟수 타입 변경(int → long)

- 자동차 대수, 시도 횟수 타입 변경: int → long

* feat(CarRaceApplication): 자동차 경기 시작 방식 변경

- 자동차 대수 대신 자동차 이름을 csv 형태로 입력
- 경기가 종료 후 우승자 출력

* refactor(CarRaceApplication): 실행 결과 메세지 출력 책임 변경

- 실행 결과 메세지 출력 책임 변경: CarRaceApplication → RaceMonitor
- RaceMonitor 내 메세지 위치 변경: PromptMessage

* feat(RaceManager): RaceManager(어플리케이션 계층) 클래스 신규

- Presentation 계층에서 수행하던 책임을 Application 계층으로 이동
- 도메인 객체를 활용하여 기능 실행
- RaceMonitor 객체는 단순 출력 기능만 수행

* docs(README.md): 기능 목록 갱신

- 신규 기능 반영에 대한 기능 목록 갱신

* feat(MoveStrategyGenerator): MoveStrategy 객체 생성용 유틸 클래스 신규

* refactor(RandomZeroAndPositiveIntegerGeneratorTest): 미사용 테스트 삭제

- RandomMoveStrategy 모킹 테스트는 RandomMoveStrategyTest에서 수행

* refactor(RandomZeroAndPositiveIntegerGenerator): Random 타입 인스턴스 변수 상수화

+++ src/main/java/study/carrace/step3/presentation/exception/IllegalCarQuantityException.java
@@ -3,7 +3,7 @@ package study.carrace.step3.presentation.exception;
 public class IllegalCarQuantityException extends RuntimeException {
     private static final String MESSAGE = "자동차 대수는 한 대 이상이여야 합니다";
 
-    public IllegalCarQuantityException(int carQuantity) {
+    public IllegalCarQuantityException(long carQuantity) {
         super(MESSAGE + ": " + carQuantity);
     }
 }
